Problem_B/requirements.txt
anthropic>=0.84.0
fastapi>=0.110.0
uvicorn>=0.27.0
pyyaml>=6.0.0
pydantic>=2.0.0
